Remove hdf5 folder and move hdf impl into db folder
